\name{hordeum}
\alias{hordeum}
\docType{data}
\title{Phylogeny of Wild Barley (Hordeum)
}
\description{
A phylogenetic tree of \emph{Hordeum} based on six nuclear DNA markers. 
}
\usage{data(hordeum)}
\format{
  The format is:
List of 17
 $ edge              : int [1:38, 1:2] 21 21 22 23 24 24 23 22 25 25 ...
 $ edge.length       : num [1:38] 11.584 0.732 2.615 3.734 4.503 ...
 $ Nnode             : int 19
 $ tip.label         : chr [1:20] "Psathyrostachys" "H.__vulgare" "H.__bulbosum" "H.__murinum" ...
 $ rate_range_MIN    : num [1:19] NA NA 0 0 0 0 0 0 0 0 ...
 $ rate_range_MAX    : num [1:19] NA NA 0.01 0.03 0.01 0.03 0.03 0.03 0.02 0.02 ...
 $ height_median     : num [1:19] 11.42 NA 8.18 4.4 8.45 ...
 $ height            : num [1:19] 11.58 NA 8.24 4.5 8.51 ...
 $ rate_95%_HPD_MIN  : num [1:19] NA NA 0 0 0 0 0 0 0 0 ...
 $ rate_95%_HPD_MAX  : num [1:19] NA NA 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 ...
 $ height_95%_HPD_MIN: num [1:19] 6.47 NA 4.48 1.86 4.54 2.82 2.34 1.78 1.02 0.29 ...
 $ height_95%_HPD_MAX: num [1:19] 16.83 NA 12.29 7.18 12.27 ...
 $ rate_median       : num [1:19] NA NA 0 0 0 0 0 0 0 0 ...
 $ height_range_MIN  : num [1:19] 2 NA 1.28 0.43 1.33 0.71 0.67 0.61 0.3 0.09 ...
 $ height_range_MAX  : num [1:19] 24.9 NA 16.6 11.3 17.4 ...
 $ rate              : num [1:19] NA NA 0 0 0 0 0 0 0 0 ...
 $ posterior         : num [1:19] 1 0.44 0.96 1 0.94 1 0.84 1 1 1 ...
 - attr(*, "class")= chr "phylo"
}

\references{
Jakob, S., C. Heibl, D. Roedder & F. Blattner. 2010. Population demography influences climatic niche evolution: evidence from South American \emph{Hordeum} species (Poaceae). \emph{Molecular Ecology} \bold{19}: 1423-1438.
}
\examples{
data(hordeum)

names(hordeum)

# plot tree
plot(hordeum)
node.support(hordeum$posterior)
}
\keyword{datasets}
