\name{phyloch-internal}
\alias{add.scale.bar2}
\alias{bigtree}
\alias{BOTHlabels2} 
\alias{bowed}
\alias{broken.segments} 
\alias{c.binary} 
\alias{degap} 
\alias{delete.nonoverlapping.seq} 
\alias{divide.angle}
\alias{extractBEASTstats}
\alias{extractSTARBEASTstats} 
\alias{extractMRBAYESstats} 
\alias{geophylo2}
\alias{identicalTips}
\alias{ladderize2}
\alias{locpie}
\alias{mafft.6.24}
\alias{match.nodelabels}
\alias{mixed.fonts.text}
\alias{mrp}
\alias{pal}
\alias{pie.at.tip}
\alias{piecoltips}
\alias{read.xml}
\alias{saturation.plot}
\alias{strwidth90}
\alias{strheight90}
\alias{tiplabels2}
\alias{treeannotator}
\alias{unresolve.clade}
\alias{write.sequin}

\title{Internal Phyloch Functions}
\description{Internal \pkg{phyloch} functions.}
\note{
These are internal functions in PHYLOCH and are not intended to be called by the user.
}
\keyword{internal}
