\name{rotate2}
\alias{rotate2}

\title{Swapping Sister Clades}

\description{
This function works exactly in the same manner as \code{\link{rotate}} in the \code{\link{ape}} package, but in addition can handle non-standard node label elements like those generated by \code{\link{read.beast}}.
}
\usage{
rotate2(phy, node, polytom = c(1, 2))
}

\arguments{
  \item{phy}{
An object of class \code{"phylo"}.
}
  \item{node}{
A vector of mode numeric or character specifying the number of the node.
}
  \item{polytom}{
A vector of mode numeric and length two specifying the two clades that should be exchanged in a polytomy.
}
}
\details{
Please refer to \code{\link{rotate}}.
}
\value{
An object of class \code{"phylo"}.
}


\author{
Christoph Heibl
}

\seealso{\code{\link{read.beast}}, \code{\link{rotate}}}
\examples{
# a phylogeny of wild barley
# --------------------------
data(hordeum)
hordeum$tip.label <- LETTERS[1:20]

# rotate phylogeny at node 27
# ---------------------------
node <- 27
hordeum.rotated <- rotate2(hordeum, node)

# plot ...
par(mfcol = c(1, 2), mar = c(0.5, 0.5, 0.5, 0.5))
# ... unmodified tree and 
# -----------------------
plot(hordeum)
nodeInfo(node = node, label = "node to rotate", angle = 160, lex = 8)
node.support(hordeum$posterior, pos = "pretty")


# ... rotated tree
# ----------------
plot(hordeum.rotated, direction = "leftward")
nodeInfo(node = node, label = "rotated node", angle = 340, lex = 6)
node.support(hordeum.rotated$posterior, pos = "pretty")
}
